First attempt at program delivery
Dig into asdf to find out traverse the list of needed fasls to load
the entire program.  Copy these into the 'objs' directory, along with
a script to load them.  The build works with sbcl and ccl, however,
with ccl, (command-line-arguments) returns the script invocation
arguments as well, so the resulting code isn't yet useful.