Initial SBCL-specific posix support
Filesystem read support using SB-POSIX.  There is also some hackery
available to allow a development lisp image to have root in its saved
uid set, which allows it to switch to an effective root uid
temporarily to allow operations to be tested that would otherwise only
be possible while root.

It's also possible to do the testing under fakeroot, which will
simulate root access, but not perform the operations as root.