Reindex broken indexes
If we get an error reading the index for a pool file, recompute the
index by scanning the chunks of the file.  Tests both the case of a
missing index, and an index that is shorter than the size of the chunk
file.  Corrupt data files will still result in an error being thrown
as will problems writing the new index out.

The index is rebuilt from scratch.  It may be possible to rebuild a
shortened index by only scanning the file from that position, but it
may not be worth it.